Add retry + 15-min TTL cache to yfinance fetchers; market_cap REAL
New ingestion/_resilience.py (stdlib only): with_retry does exponential
backoff with jitter and logs retry.attempt; TTLCache holds successful
results for 15 minutes. Both yfinance fetchers check the cache first
(cache hits keep their original fetched_at) and retry the network call.
tests/conftest.py clears the caches around every test.

Also aligns the fundamentals schema with the dataclass: market_cap is
now REAL, matching float | None.

+++ storage/store.py
@@ -49,7 +49,7 @@ CREATE TABLE IF NOT EXISTS fundamentals (
     name           TEXT,
     sector         TEXT,
     industry       TEXT,
-    market_cap     INTEGER,
+    market_cap     REAL,
     trailing_pe    REAL,
     forward_pe     REAL,
     price_to_book  REAL,
